fixed data type

+++ dbt_subprojects/solana/models/_sector/dex/meteora/meteora_v4_solana_base_trades.sql
@@ -71,8 +71,8 @@ select
 , sd.block_time 
 , sd.block_slot 
 , sd.trade_source
-, cast(evt.token_out_amount_raw as double) as token_bought_amount_raw
-, cast(evt.token_in_amount_raw as double) as token_sold_amount_raw
+, cast(evt.token_out_amount_raw as uint256) as token_bought_amount_raw
+, cast(evt.token_in_amount_raw as uint256) as token_sold_amount_raw
 , cast (null as double) as fee_tier
 -- , evt.total_fees_raw
 , case when evt.trade_direction =0 then sd.token_a else sd.token_b end as token_sold_mint_address
